apps/feedback-viewer/app/backend/requirements.txt
fastapi
uvicorn
azure-cosmos
azure-identity
pydantic
python-dotenv